Enhancement: Extract factory and rule sets
