+++ .php_cs
@@ -11,7 +11,7 @@ the LICENSE file that was distributed with this source code.
 @link https://github.com/localheinz/php-cs-fixer-config
 EOF;
 
-$config = new Config\Php56($header);
+$config = Config\Factory::fromRuleSet(new Config\RuleSet\Php56($header));
 
 $config->getFinder()->in(__DIR__);
